backend/requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── LangGraph orchestration (agentic pipeline) ───────────────────────────────
langgraph>=1.0.7
langchain-core>=1.2.24

# ── Vector store & retrieval ──────────────────────────────────────────────────
qdrant-client>=1.9.0

# ── HTTP client (Azure Embedding API + Qdrant) ───────────────────────────────
httpx>=0.27.0

# ── LLM provider (Gemini) ────────────────────────────────────────────────────
google-genai>=1.70.0

# ── Language detection ────────────────────────────────────────────────────────
langdetect
